add import/export files to main.js
